[NF] #Added sorting method by post relevance
